Add an abstract mother class for depth-first searches.
